Plot MI Network with Community Coloring

Visualizes a subset of an MI network with community coloring, labeling top nodes from each community.

Usage

plot_mi_network(
  g_consensus,
  large_communities,
  sample_size = 100,
  top_n_per_comm = 5,
  title = "Network Communities"
)

Arguments

g_consensus

An igraph object representing the consensus graph.

large_communities

A list of communities, where each element contains the nodes of a community.

sample_size

Integer, number of nodes to sample from the graph (default: 100).

top_n_per_comm

Integer, number of top nodes per community to label (default: 5).

title

Character, title of the plot (default: "Network Communities").

Value

A ggraph plot object.